Gjorde UserInterface.confirm återanvändningsbar, genom att lägga till String question.
La till bekräftelse på factory reset

Skapar nu databas vid programstart om det inte redan finns en.

+++ src/com/company/Management.java
@@ -164,7 +164,7 @@ public class Management implements Serializable {
 
             if ((roomChoice == roomsAvailable)) {
                 roomDetails(roomChoice);
-                if (USER_INTERFACE.confirm()) {
+                if (USER_INTERFACE.confirm("Would you like to book this room?")) {
 
                     int numberOfNights = USER_INTERFACE.enterInteger("number of nights to book", 1, 14);
 
@@ -176,8 +176,8 @@ public class Management implements Serializable {
                     int customerID2 = ResultClass.ifInputMatches(sqlStatement, query, "customer_id");
 
                     if ((customerID2 == customerID)) {
-                        System.out.println("You allready have a booked room, do you want to check out?");
-                        if (USER_INTERFACE.confirm()) {
+                        System.out.println("You already have a booked room, do you want to check out?");
+                        if (USER_INTERFACE.confirm("Would you like to book this room? (Y/N)")) {
                             checkOutWithBill(customerID);
                         }
                     } else {
@@ -289,45 +289,56 @@ public class Management implements Serializable {
     }
 
     public void setup() throws SQLException, IOException {
-        sqlStatement = Run.getSqlStatement();
-        List<Room> newFile = Room.rooms();
-        DataHandler.writeToFile(newFile, "Rooms.txt");
+        if (USER_INTERFACE.confirm("Are you sure that you want to Factory Reset?")) {
 
-        //Creating database
-        String dropDatabase = "DROP DATABASE hotel_booking_system;";
-        String createDatabase = "CREATE DATABASE hotel_booking_system;";
-        String useDatabase = "USE hotel_booking_system";
-        String createCustomerTable = "CREATE TABLE customer (id INT NOT NULL PRIMARY KEY AUTO_INCREMENT, customer_name VARCHAR(50) NOT NULL, contact_number VARCHAR(50));";
-        String createRoomTable = "CREATE TABLE room (roomNumber INT NOT NULL PRIMARY KEY AUTO_INCREMENT, typeOfRoom INT NOT NULL, room_Description VARCHAR(250), price_per_night INT NOT NULL, wifi BIT DEFAULT 1, tv VARCHAR(50) NOT NULL, aircondition BIT NOT NULL, smoking BIT DEFAULT 0);";
-        String createRoomBookingTable = "CREATE TABLE roomBooking (id INT NOT NULL PRIMARY KEY AUTO_INCREMENT, roomNumber INT NOT NULL, customer_id INT DEFAULT NULL, roomAvailable BIT DEFAULT 1, checkInDate DATETIME DEFAULT NULL, checkOutDate DATETIME DEFAULT NULL, FOREIGN KEY (customer_id) REFERENCES customer(id) ON DELETE CASCADE,FOREIGN KEY (roomnumber) REFERENCES room(roomNumber) ON DELETE CASCADE);";
-        String createFoodOrderTable = "CREATE TABLE foodOrder (id INT NOT NULL PRIMARY KEY AUTO_INCREMENT, food_id INT NOT NULL, customer_id INT, FOREIGN KEY (customer_id) REFERENCES customer(id) ON DELETE CASCADE);";
-        String createViewAvailableRooms = "CREATE VIEW availableRooms AS SELECT DISTINCT room.roomNumber, typeOfRoom FROM room JOIN roomBooking ON roomBooking.roomNumber = room.roomNumber WHERE roomAvailable = 1;";
-        String createViewBookedRooms = "CREATE VIEW bookedroom AS SELECT DISTINCT customer_id, typeOfroom, price_per_night FROM room JOIN roombooking ON room.roomnumber = roombooking.roomnumber;";
+            sqlStatement = Run.getSqlStatement();
+            List<Room> newFile = Room.rooms();
+            DataHandler.writeToFile(newFile, "Rooms.txt");
 
-        //Creating rooms
-        String createStandardSingle = "INSERT INTO room (typeOfRoom, price_per_night, wifi, tv, aircondition) VALUES (1, 450, 0, 'big old tv', 1);";
-        String createStandardDouble = "INSERT INTO room (typeOfRoom, price_per_night, wifi, tv, aircondition)VALUES (2, 600, 0, 'big old tv', 1)";
-        String createDeluxeSingle = "INSERT INTO room (typeOfRoom, price_per_night, wifi, tv, aircondition)VALUES (3, 850, 1, 'flat screen', 1);";
-        String createDeluxeDouble = "INSERT INTO room (typeOfRoom, price_per_night, wifi, tv, aircondition)VALUES (4, 1250, 1, 'flat screen', 1);";
+            //#region Creating database
+            String dropDatabase = "DROP DATABASE hotel_booking_system;";
+            String createDatabase = "CREATE DATABASE hotel_booking_system;";
+            String useDatabase = "USE hotel_booking_system";
+            String createCustomerTable = "CREATE TABLE customer (id INT NOT NULL PRIMARY KEY AUTO_INCREMENT, customer_name VARCHAR(50) NOT NULL, contact_number VARCHAR(50));";
+            String createRoomTable = "CREATE TABLE room (roomNumber INT NOT NULL PRIMARY KEY AUTO_INCREMENT, typeOfRoom INT NOT NULL, room_Description VARCHAR(250), price_per_night INT NOT NULL, wifi BIT DEFAULT 1, tv VARCHAR(50) NOT NULL, aircondition BIT NOT NULL, smoking BIT DEFAULT 0);";
+            String createRoomBookingTable = "CREATE TABLE roomBooking (id INT NOT NULL PRIMARY KEY AUTO_INCREMENT, roomNumber INT NOT NULL, customer_id INT DEFAULT NULL, roomAvailable BIT DEFAULT 1, checkInDate DATETIME DEFAULT NULL, checkOutDate DATETIME DEFAULT NULL, FOREIGN KEY (customer_id) REFERENCES customer(id) ON DELETE CASCADE,FOREIGN KEY (roomnumber) REFERENCES room(roomNumber) ON DELETE CASCADE);";
+            String createFoodOrderTable = "CREATE TABLE foodOrder (id INT NOT NULL PRIMARY KEY AUTO_INCREMENT, food_id INT NOT NULL, customer_id INT, FOREIGN KEY (customer_id) REFERENCES customer(id) ON DELETE CASCADE);";
+            String createViewAvailableRooms = "CREATE VIEW availableRooms AS SELECT DISTINCT room.roomNumber, typeOfRoom FROM room JOIN roomBooking ON roomBooking.roomNumber = room.roomNumber WHERE roomAvailable = 1;";
+            String createViewBookedRooms = "CREATE VIEW bookedroom AS SELECT DISTINCT customer_id, typeOfroom, price_per_night FROM room JOIN roombooking ON room.roomnumber = roombooking.roomnumber;";
+            //#endregion
 
-        ResultClass.setupResult(sqlStatement, dropDatabase);
-        ResultClass.setupResult(sqlStatement, createDatabase);
-        ResultClass.setupResult(sqlStatement, useDatabase);
-        ResultClass.setupResult(sqlStatement, createCustomerTable);
-        ResultClass.setupResult(sqlStatement, createRoomTable);
-        ResultClass.setupResult(sqlStatement, createRoomBookingTable);
-        ResultClass.setupResult(sqlStatement, createFoodOrderTable);
-        ResultClass.setupResult(sqlStatement, createViewAvailableRooms);
-        ResultClass.setupResult(sqlStatement, createViewBookedRooms);
-        ResultClass.setupResult(sqlStatement, createStandardSingle);
-        ResultClass.setupResult(sqlStatement, createStandardDouble);
-        ResultClass.setupResult(sqlStatement, createDeluxeSingle);
-        ResultClass.setupResult(sqlStatement, createDeluxeDouble);
+            //#region Creating rooms
+            String createStandardSingle = "INSERT INTO room (typeOfRoom, price_per_night, wifi, tv, aircondition) VALUES (1, 450, 0, 'big old tv', 1);";
+            String createStandardDouble = "INSERT INTO room (typeOfRoom, price_per_night, wifi, tv, aircondition)VALUES (2, 600, 0, 'big old tv', 1)";
+            String createDeluxeSingle = "INSERT INTO room (typeOfRoom, price_per_night, wifi, tv, aircondition)VALUES (3, 850, 1, 'flat screen', 1);";
+            String createDeluxeDouble = "INSERT INTO room (typeOfRoom, price_per_night, wifi, tv, aircondition)VALUES (4, 1250, 1, 'flat screen', 1);";
+            //#endregion
 
-        //Inserting rooms into booking system
-        for (int i = 1; i < 5; i++) {
-            String createRoomNumbers = "INSERT INTO roombooking (roomnumber) VALUES (" + i + ");";
-            ResultClass.setupResult(sqlStatement, createRoomNumbers);
+            //#region Sending queries to mySql
+            ResultClass.setupResult(sqlStatement, dropDatabase);
+            ResultClass.setupResult(sqlStatement, createDatabase);
+            ResultClass.setupResult(sqlStatement, useDatabase);
+            ResultClass.setupResult(sqlStatement, createCustomerTable);
+            ResultClass.setupResult(sqlStatement, createRoomTable);
+            ResultClass.setupResult(sqlStatement, createRoomBookingTable);
+            ResultClass.setupResult(sqlStatement, createFoodOrderTable);
+            ResultClass.setupResult(sqlStatement, createViewAvailableRooms);
+            ResultClass.setupResult(sqlStatement, createViewBookedRooms);
+            ResultClass.setupResult(sqlStatement, createStandardSingle);
+            ResultClass.setupResult(sqlStatement, createStandardDouble);
+            ResultClass.setupResult(sqlStatement, createDeluxeSingle);
+            ResultClass.setupResult(sqlStatement, createDeluxeDouble);
+            //#endregion
+
+            //Inserting rooms into booking system
+            for (int i = 1; i < 5; i++) {
+                String createRoomNumbers = "INSERT INTO roombooking (roomnumber) VALUES (" + i + ");";
+                ResultClass.setupResult(sqlStatement, createRoomNumbers);
+            }
+            System.out.println("Factory Reset - Complete\n" +
+                    "Database has been set up from scratch!\n");
+        } else {
+            System.out.println("Factory Reset - Cancelled\n");
         }
     }
 }
